add a crypto function to validate incoming data
Github will send a hashed header based on their POST body and
our secret.  We need to hash the body with our secret, and compare
that to their given hmac header. We need to use a secure compare
to prevent any potential timing attacks based on the length of
these values.