add a check to verify the hashed body is the right length
We check that the two buffers are the same length before comparing,
and we verify that the original hashed header is 64 characters
before converting to a buffer.

The buffer truncates odd numbered string values before creating a
buffer. So, if we receive a value that is off by one character, this
check will reject that value. Whereas otherwise, that value would
succeed because Buffer would truncate the last incorrect value.